Refactored output writers for text or html.
Replaced the eval approach by OO with inheritance, with the HTML output writer
being a subclass of Output::Text. Basically this is the result of playing
around with Perl's OO mechanisms. There is no noteworthy new functionality, at
best I made the code a bit cleaner and didn't break anything.